requirements.txt
torch
transformers==4.44.2
datasets==2.21.0
accelerate==0.34.2
tokenizers==0.19.1
bitsandbytes
peft
albumentations
scipy==1.13.1
numpy==1.24.4
pillow
evaluate
jiwer
scikit-learn
pandas
xxhash
huggingface-hub
sentencepiece
protobuf